Create README.md
Project Overview:
In this project, I have applied the skills to build a pipeline to process real-world, user-supplied images.
The data consists of 8351 dog images. The images are sorted into 133 directories, each directory contains only images of a single dog breed.
I use Keras to build a convolutional neural network to recognize dog breeds from images. Given an image of a dog, the model will identify the canine’s breed. If supplied an image of a human, the code will identify the resembling dog breed.